% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_attgt_gmm.R
\name{process_attgt_gmm}
\alias{process_attgt_gmm}
\title{Process Results from \code{\link[=compute.att_gt]{compute.att_gt()}}}
\usage{
process_attgt_gmm(attgt.list)
}
\arguments{
\item{attgt.list}{list of results from \code{\link[=compute.att_gt]{compute.att_gt()}}}
}
\value{
list with elements:
\item{group}{which group a set of results belongs to}
\item{tt}{which time period a set of results belongs to}
\item{att}{the group time average treatment effect}
}
\description{
Process Results from \code{\link[=compute.att_gt]{compute.att_gt()}}
}
